+++ docs/changes.rst
@@ -13,6 +13,7 @@ Not yet released.
 
 * Rendering of release notes on GitHub.
 * Listing of categorized projects.
+* Translating a language inside a category.
 
 **Compatibility**